+++ apps/coder/src/pages/ChatPage-MCPTest.tsx
@@ -1,0 +1,9 @@
+import { useApiKeyContext } from "@/providers/ApiKeyProvider";
+import { useOpenAgent } from "@openagents/core";
+import { AgentChat } from "@openagents/ui";
+
+export default function ChatPage() {
+  const { apiKeys } = useApiKeyContext();
+  const agent = useOpenAgent('coder')
+  return <AgentChat agent={agent} githubToken={apiKeys.github} />
+}
